JS testing added, added one test for testing hello world
