+++ jam_scene_backend/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "npx jest",
     "start": "node server.js"
   },
   "author": "",
@@ -17,6 +17,8 @@
   "devDependencies": {
     "eslint": "^8.19.0",
     "eslint-config-prettier": "^8.5.0",
-    "prettier": "2.7.1"
+    "jest": "^28.1.2",
+    "prettier": "2.7.1",
+    "supertest": "^6.2.4"
   }
 }
